feat: Implementando o GrupoPagamentoDTO e adicionando validação de transações existentes em uma conta antes de remove-la

+++ app-financeiro-back-end/src/main/java/bcd/appfinanceirobackend/service/ContaService.java
@@ -6,6 +6,7 @@ import bcd.appfinanceirobackend.exception.ResourceNotFoundException;
 import bcd.appfinanceirobackend.model.Conta;
 import bcd.appfinanceirobackend.model.Usuario;
 import bcd.appfinanceirobackend.repository.ContaRepository;
+import bcd.appfinanceirobackend.repository.TransacaoRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -17,8 +18,13 @@ import java.util.UUID;
 public class ContaService {
 
     private final ContaRepository contaRepository;
+    private final TransacaoRepository transacaoRepository;
 
-    public ContaService (ContaRepository contaRepository) {this.contaRepository = contaRepository;}
+    public ContaService (ContaRepository contaRepository,
+                         TransacaoRepository transacaoRepository) {
+        this.contaRepository = contaRepository;
+        this.transacaoRepository = transacaoRepository;
+    }
 
     public List<ContaResponseDTO> listarPorUsuario(Usuario usuarioAutenticado) {
         return contaRepository.findByUsuarioId(usuarioAutenticado.getId())
@@ -48,7 +54,16 @@ public class ContaService {
         if (!conta.getUsuario().getId().equals(usuario.getId())) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acesso negado a esta conta");
         }
-        contaRepository.delete(conta);
+        /* Apesar de o banco possuir ON DELETE CASCADE para a relação entre contas e transações,
+         * a regra de negócio adotada no service bloqueia a exclusão de contas com transações vinculadas
+         * para evitar perda acidental de histórico financeiro. O cascade permanece apenas como proteção estrutural
+         * do banco, mas o fluxo da aplicação impede a remoção destrutiva.
+        */
+         if (transacaoRepository.existsByContaId(conta.getId())) {
+             throw new ResponseStatusException(HttpStatus.CONFLICT,
+                     "Não é possível remover uma conta com transações vinculadas");
+         }
+         contaRepository.delete(conta);
     }
 
     public ContaResponseDTO toResponse(Conta conta) {
